pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kukulkan-trade"
version = "0.1.0"
description = "Educational trading bot running 2 portfolio strategies on Alpaca paper trading"
requires-python = ">=3.11"
dependencies = [
    # Alpaca API
    "alpaca-py>=0.30.0",
    # AI
    "anthropic>=0.39.0",
    # Vector store client (server runs in Docker)
    "chromadb-client>=0.5.0",
    # News data
    "finnhub-python>=2.4.0",
    # Data & analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    "yfinance>=0.2.36",
    # Macro data
    "fredapi>=0.5.2",
    # Market calendar
    "exchange-calendars>=4.5",
    # Scheduling
    "APScheduler>=3.10.4",
    # Telegram
    "python-telegram-bot>=21.0",
    # Config & logging
    "pydantic-settings>=2.4.0",
    "structlog>=24.4.0",
    # Database
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that hit external APIs",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
